requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Environment configuration
django-environ==0.11.2

# Authentication and security
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
argon2-cffi==23.1.0
cryptography==41.0.7

# Validation and utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Logging and monitoring
structlog==23.2.0